code cleanup and initial setup for this branch
