+++ components/layout.tsx
@@ -2,7 +2,8 @@ import MyHead from "./sections/myHead";
 import NavBar from "./sections/navbar";
 import Footer from "./sections/footer";
 import { ReactElement } from "react";
-import { navLinks, getPageInfo } from "@/public/utils/data";
+import { navLinks } from "@/public/utils/data";
+import { getPageInfo } from "@/public/utils/helperFunctions";
 
 export default function Layout({children}:{children:ReactElement}) {
     const pageInfo = getPageInfo();
