+++ public/utils/data.tsx
@@ -1,5 +1,3 @@
-import { useRouter } from "next/router";
-
 export const navLinks = [
     {
         pageTitle: "Home",
@@ -59,14 +57,3 @@ export const webLinks = {
     facebook: "https://www.facebook.com/profile.php?id=100077784492698",
     youtube: "https://www.youtube.com/@hbcustemer5734",
 }
-
-export function getPageInfo() {
-    const router = useRouter();
-    const pathname = router.pathname;
-
-    const filtered = navLinks.filter(pageInfo => pageInfo.path === pathname);
-
-    const pageInfo = filtered[0];
-
-    return pageInfo;
-  }
